requirements.txt
torch==2.2.0
torchvision==0.17.0
pandas==2.2.0
PyYAML==6.0.1
Pillow==10.2.0
scikit-learn==1.4.0
transformers==4.38.0
scikit-learn==1.4.0
tqdm==4.66.0
shap==0.45.0
requests==2.31.0
